+++ design/spectre.js
@@ -1,0 +1,29 @@
+import React from 'react';
+
+//elements
+export * as Typography from '@react-spectre/typography/src';
+export * as Table from '@react-spectre/table/src';
+export * as Button from '@react-spectre/button/src';
+export * as Form from '@react-spectre/form/src';
+export * as Icon from '@react-spectre/icon/src';
+export * as Label from '@react-spectre/label/src';
+export * as Media from '@react-spectre/media/src';
+
+//layout
+export * as Layout from '@react-spectre/layout/src';
+export * as Grid from '@react-spectre/grid/src';
+export * as Navbar from '@react-spectre/navbar/src';
+
+//components
+export * as Accordion from '@react-spectre/accordion/src';
+export * as Autocomplete from '@react-spectre/autocomplete/src';
+export * as Avatar from '@react-spectre/avatar/src';
+export * as Badge from '@react-spectre/badge/src';
+export * as Bar from '@react-spectre/bar/src';
+export * as Toast from '@react-spectre/toast/src';
+export * as Card from '../lib/vendor/react-spectre-master/packages/card/src/index';
+export * as Panel from '../lib/vendor/react-spectre-master/packages/panel/src/index';
+export * as Step from '../lib/vendor/react-spectre-master/packages/step/src/index';
+export * as Popover from '../lib/vendor/react-spectre-master/packages/popover/src/index';
+export * as Tile from '../lib/vendor/react-spectre-master/packages/tile/src/index';
+