+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## 2.2.0 (2019-04-01)
+
+- **Feat**
+  - Added React.lazy support
+  - Added React.Suspense support
+- **Fix**
+  - Named class support via options
+  - passing state and props to class render component via options
+- **Chore**
+  
+## 2.1.0 (2019-04-01)
+
+- **Feat**
+- **Fix**
+  - Improved inline function support
+  - Improved debugging support
+  - Automated parameter detection
+- **Chore**
+  - Update Documentation
+
 ## 2.0.0 (2019-03-29)
 
 - **Feat**